// server/client/models/changeUserRoleControllerHandleBodyRole.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * IFAL Arapiraca Backend API
 * API para gerenciamento de problemas de infraestrutura do IFAL Arapiraca
 * OpenAPI spec version: 1.0.0
 */

export type ChangeUserRoleControllerHandleBodyRole =
  (typeof ChangeUserRoleControllerHandleBodyRole)[keyof typeof ChangeUserRoleControllerHandleBodyRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ChangeUserRoleControllerHandleBodyRole = {
  REPORTER: "REPORTER",
  MANAGER: "MANAGER",
  DIRECTOR: "DIRECTOR",
  ADMIN: "ADMIN",
} as const;